requirements.txt
langchain==1.2.15
langchain-openai==1.1.12
langgraph==1.1.5
openai==2.30.0
python-dotenv==1.2.2
